Ver 2.0.1.29  SuperReview version 2 fixes for Steel Piece work Files modified - [PartFASteelStandard,ReviewSuper,Job1111,.wmproject.properties,wmProperties.js]

+++ services/Job1111/src/com/psabackoffice/job1111/models/query/GetActivitiesPendingReviewSteelActivitiesResponse.java
@@ -24,7 +24,7 @@ public class GetActivitiesPendingReviewSteelActivitiesResponse implements Serial
     private String activityType;
 
     @ColumnAlias("ActivityTypeID")
-    private Integer activityTypeId;
+    private Short activityTypeId;
 
     @ColumnAlias("ActivityID")
     private Integer activityId;
@@ -71,11 +71,11 @@ public class GetActivitiesPendingReviewSteelActivitiesResponse implements Serial
         this.activityType = activityType;
     }
 
-    public Integer getActivityTypeId() {
+    public Short getActivityTypeId() {
         return this.activityTypeId;
     }
 
-    public void setActivityTypeId(Integer activityTypeId) {
+    public void setActivityTypeId(Short activityTypeId) {
         this.activityTypeId = activityTypeId;
     }
